internal/metamorphic: refactor to expose package interface
Refactor the metamorphic tests to expose a package interface that allows other
packages to invoke the metamorphic tests. Future work will build on this to
move the package interface into a non-internal package (so that it may be
invoked from the Cockroach repo) and expose additional options for configuring
the runs.

Eventually, I'd like to invoke the metamorphic tests with some runs using the
encryption-at-rest filesystem.
